updated UnCountable items case

+++ src/main/java/org/shoppingcart/cartitem/CartItem.java
@@ -7,10 +7,10 @@ import java.math.BigDecimal;
 public class CartItem {
     private Product product;
     private BigDecimal cost;
-    private Integer quantity;
-    private Integer weight;
+    private BigDecimal quantity;
+    private BigDecimal weight;
 
-    public CartItem(Product product, Integer howMuch) {
+    public CartItem(Product product, BigDecimal howMuch) {
         this.product = product;
         switch (product.getType()){
             case COUNTABLE -> this.quantity = howMuch;
@@ -23,26 +23,26 @@ public class CartItem {
         return product;
     }
 
-    public Integer getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public Integer getWeight() {
+    public BigDecimal getWeight() {
         return weight;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
 
-    public void setWeight(Integer weight) {
+    public void setWeight(BigDecimal weight) {
         this.weight = weight;
     }
 
     public BigDecimal getCost() {
         switch(getProduct().getType()){
-            case COUNTABLE -> cost = getProduct().getPrice().multiply(BigDecimal.valueOf(getQuantity()));
-            case NON_COUNTABLE -> cost = getProduct().getPrice().multiply(BigDecimal.valueOf(getWeight()));
+            case COUNTABLE -> cost = getProduct().getPrice().multiply(getQuantity());
+            case NON_COUNTABLE -> cost = getProduct().getPrice().multiply(getWeight());
         }
         return cost;
     }
